Split QueryBuilder.unit.spec.tsx into QueryBuilder.beforeunload-events.unit.spec.tsx, QueryBuilder.unsaved-changes-warning.unit.spec.tsx & test-utils.tsx
